requirements.txt
flask
flask-sqlalchemy
python-dotenv
flask-migrate
requests